[pythonic] readability-redundant-string-init, readability-redundant-inline-specifier...
modernize-use-nullptr,
readability-redundant-control-flow,
readability-redundant-casting,
readability-redundant-member-init,
readability-redundant-typename,
readability-use-anyofallof,
readability-use-std-min-max,
readability-redundant-function-ptr-dereference,
readability-static-definition-in-anonymous-namespace,

+++ pythran/pythonic/numpy/random/logseries.hpp
@@ -48,7 +48,7 @@ namespace numpy
         U = std::uniform_real_distribution<double>{0., 1.}(details::generator);
         q = -expm1(r * U);
         if (V <= q * q) {
-          result = (double)floor(1 + log(V) / log(q));
+          result = floor(1 + log(V) / log(q));
           if ((result < 1) || (V == 0.0)) {
             continue;
           } else {
